requirements.txt
factory-boy==3.3.3
playwright==1.53.0
pydantic==2.11.7
pytest==8.4.1
pytest-playwright==0.7.0
pyyaml==6.0.2
ruff==0.12.4
